+++ src/main/java/web/service/UserService.java
@@ -7,11 +7,11 @@ import java.util.List;
 public interface UserService {
     List<User> getAllUsers();
 
-    User getUser(int id);
+    User getUser(long id);
 
     void createUser(User user);
 
     void updateUser(User user);
 
-    void deleteUser(int id);
+    void deleteUser(long id);
 }
